+++ lib/dotenv_validator/errors.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module DotenvValidator # :nodoc: all
+  class SampleFileNotFoundError < StandardError
+  end
+end